+++ product-api/src/main/kotlin/com/example/productapi/product/api/exceptionhandler/ProductExceptionHandler.kt
@@ -1,4 +1,4 @@
-package com.example.productapi.product.exceptionhandler
+package com.example.productapi.product.api.exceptionhandler
 
 import lombok.extern.slf4j.Slf4j
 import org.springframework.http.HttpStatus
