+++ product-api/src/main/kotlin/com/example/productapi/product/api/request/ProductEditRequest.kt
@@ -1,4 +1,4 @@
-package com.example.productapi.product.controller.request
+package com.example.productapi.product.api.request
 
 import com.example.productdomain.product.application.dto.ProductEditInput
 import com.example.productdomain.product.domain.ProductStatus
